Adds eight_queens solution
= My approach
Iterated over every single square in the board and recursively
counted number of ways to place a queen, by maintaining a vector
that stops growing once it reaches size 8 (base case).

Did not get this working.

= Solutions
A better approach was to recurse on the current row to place
a queen in. This avoids considering cells in the same row
by construction. Could have also done something similar on cols,
both are equivalent.

Then, we can only place a queen if the cell is not already covered
by an obstacle, or it is out of range of the other already placed
queens.