Solves "Why did the Cow cross the road II"
= My solution
Tried to use a combination of counters and start/end for certain letters.
A LOT of tunnel vision on iterating in the order of the input string.

I additionally tried to come up with an approach that extends the
idea to the leetcode problem "valid parens", but this also led me nowhere.
I didn't have an idea on how to proceed from there.

= Solution
Use brute force, it is a "complete search" problem for a reason.
Iterate over all possible cow pairs, and check if their entry/exit
paths intersect.

i.e.,
           start      end
                cow 2
              ---------
              |       |

          |      |
          --------
            cow 1
       start     end

Since we are iterating over all pairs, this is complete search,
just count such instances.