Solves USACO 2022 Problem 1: Sleeping in Class
= My approach
Implemented a recursive solution that tried merging all adjacent values
and minimises the branch that had the least merges.

= Solution
Missed a critical insight.

Arrays that can be successfully merged, by consequence of the adjacent
nature of the merge operation, have implicit "dividers" that can be
placed between array elements. The sum of elements between any two
dividers (or the array start/end) will be equal, implying the number
of operations to be # dividers + 1.

Dividers are obtained by splitting the total sum into buckets of
integral capacity.