Solves CF 863 Problem B: Kayaking
= My solution
Tried a recursive implementation, which was incorrect and very inefficient.
The idea was, to maintain state of how many skips have been done,
if we haven't reached the max skip capacity, compute the answer if we were
to skip the current person. Also compute the cost of taking the person,
calculated in two cases depending on whether this is the first person for
a kayak or the partner to a 1/2 filled kayak. Overly complicated and took a
while to even implement.

= Solution
Instead, just minimise over all possible distinct pairs of people to skip.
Then we solve the related problem of determining the minimum total instability
after excluding two people i, j. Use a greedy approach after sorting (note
that sorting is done before we consider all possible skips)

O(n^2) and for each quadratic factor we perform a linear sweep to compute
sum abs(p[i] - p[i+1]).