Solves problem USACO 2019 Problem 2: Where am I?
= My solution
Find the smallest such k that satisfies the "whereami" condition, i.e.,
given any substring of size k, this can uniquely identify the relative
position in the entire string. Binary search optimal, linear search possible
with the constraints. Had to be careful to also account for the case that
all chars in the string are unique, so k = 0 was possible.