Solves USACO 2018 Problem 3: Out of Place
= My solution
Failed. Completely. Tried to figure out which one needed to be
swapped and to figure out the distance to which I swap the
correct position with.

= Solution
Count the number of differences between the sorted and the unsorted
lists. There can either be 0 or 2 or more differences, so we use
max(0, differences - 1).