Solves CF 2172 Problem M: Maximum Distance To Port
First graph problem in a while!

We have a connected graph where a node is a city, which manufactures
products of type t_i, where 1 <= i <= k.

Our goal is to find the largest shortest distance of all such t_i 's
to the port city (labelled 1).

= My solution
Initially thought of running multiple BFS starting at each non-port
city with destination as the port city. The idea was to maximise
over the distance for the product type (as we iterate through all
cities) for all product types.

It took a dinner break to realise that since edges are bidirectional,
we can oneshot the entire dist calculation process by starting from
dest and traversing the remainder of the graph.